refactor(ManifoldRuntime): extract SessionToolDispatchBinder seam
Move the per-turn tool-dispatch wiring (advertise / register / unregister) out
of ConversationTurnExecutor into a named SessionToolDispatchBinder collaborator.
The three methods (advertisedToolDefinitions, registerSessionToolExecutors,
unregisterSessionToolExecutors) move verbatim; the executor delegates. The
register-before-enqueue ordering (#1606) is unchanged — the executor still
advertises, then registers, then calls enqueueAsync, and unregisters once the
stream drains. Behavior-preserving — goldens diff clean; SessionToolSource
dispatch + ConversationRuntime + handoff suites all green (sabotage-verified the
register seam reproduces the #1606 unknownTool failure when broken).
