+++ Sources/ManifoldRuntime/Services/ConversationTurnExecutor.swift
@@ -35,6 +35,9 @@ struct ConversationTurnExecutor: Sendable {
     /// ``RuntimeBindingsBox`` for the rationale (host-app demo swaps per
     /// scenario card; full runtime rebuild would be overkill).
     private let bindings: RuntimeBindingsBox
+    /// Per-turn session-tool advertise/register/unregister seam. See
+    /// ``SessionToolDispatchBinder`` and #1606 (register-before-enqueue).
+    private let toolDispatch: SessionToolDispatchBinder
 
     init(
         persistence: ConversationPersistencePort,
@@ -74,6 +77,7 @@ struct ConversationTurnExecutor: Sendable {
         self.hostTurnContextProvider = hostTurnContextProvider
         self.legacyTurnContextProvider = turnContextProvider
         self.bindings = bindings
+        self.toolDispatch = SessionToolDispatchBinder(inferenceService: inferenceService)
     }
 
     // MARK: Send flow
@@ -746,7 +750,7 @@ struct ConversationTurnExecutor: Sendable {
         // tool but limiting which names go on the wire works without
         // unregistering the executor. Fetched on the main actor because the
         // registry and InferenceService accessors are both MainActor-isolated.
-        let advertisedTools: [ToolDefinition] = await readAdvertisedToolDefinitions(
+        let advertisedTools: [ToolDefinition] = await toolDispatch.advertisedToolDefinitions(
             sessionToolSources: turnSessionToolSources,
             sessionRecord: sessionRecord
         )
@@ -756,7 +760,7 @@ struct ConversationTurnExecutor: Sendable {
         // dispatch loop finds no registry executor and returns `unknownTool`
         // (#1606). Registered after `advertisedTools` is computed so advertising
         // is unaffected; unregistered once the stream fully drains below.
-        let registeredSessionToolNames = await registerSessionToolExecutors(
+        let registeredSessionToolNames = await toolDispatch.registerSessionToolExecutors(
             sources: turnSessionToolSources,
             sessionRecord: sessionRecord
         )
@@ -777,7 +781,7 @@ struct ConversationTurnExecutor: Sendable {
                 requestGroupID: sessionID
             )
         } catch {
-            await unregisterSessionToolExecutors(registeredSessionToolNames)
+            await toolDispatch.unregisterSessionToolExecutors(registeredSessionToolNames)
             let conversationError = ConversationError.inference(error)
             emit(.errorRaised(conversationError))
             await completeOutcome(
@@ -930,7 +934,7 @@ struct ConversationTurnExecutor: Sendable {
         // stale ``ChatSession`` binding into the next turn (#1606). All
         // remaining exit paths below are post-stream, so this is the single
         // cleanup point alongside the enqueue-failure path above.
-        await unregisterSessionToolExecutors(registeredSessionToolNames)
+        await toolDispatch.unregisterSessionToolExecutors(registeredSessionToolNames)
 
         // Flush remaining buffered tokens (normal end, error, or cancellation).
         if let batch = batcher.flush(now: ContinuousClock.now) {
@@ -1583,133 +1587,6 @@ struct ConversationTurnExecutor: Sendable {
     private func isCancelled(handle: ConversationStreamHandle) async -> Bool {
         if Task.isCancelled { return true }
         return await registry.isCancelled(handle)
-    }
-
-    private func readAdvertisedToolDefinitions(
-        sessionToolSources: [any SessionToolSource],
-        sessionRecord: ChatSession?
-    ) async -> [ToolDefinition] {
-        // Base registry definitions — host-installed tools the executor
-        // already advertises today. Read from MainActor.
-        let registryDefinitions: [ToolDefinition] = await MainActor.run {
-            inferenceService.toolRegistry?.advertisedDefinitions ?? []
-        }
-
-        // Per-session source contributions. Each source's
-        // `toolDefinitions(for:)` returns the definitions it owns for this
-        // session — e.g. ``HandoffToolSource`` synthesising one
-        // `transfer_to_<name>` per non-active agent. Skipped when we don't
-        // have a session record to scope the query.
-        var sourceDefinitions: [ToolDefinition] = []
-        if let sessionRecord {
-            for source in sessionToolSources {
-                let defs = await source.toolDefinitions(for: sessionRecord)
-                sourceDefinitions.append(contentsOf: defs)
-            }
-        }
-
-        // Union — sources can override the registry's view of a tool of
-        // the same name by appearing later, but in practice the namespaces
-        // don't overlap (registry: bespoke executors; sources: synthesised
-        // tools). De-dupe by name to keep the wire shape stable.
-        var unioned: [ToolDefinition] = registryDefinitions
-        var seen = Set(registryDefinitions.map(\.name))
-        for def in sourceDefinitions where !seen.contains(def.name) {
-            unioned.append(def)
-            seen.insert(def.name)
-        }
-
-        // Apply allowed-tools intersection across sources. A `nil` return
-        // from `allowedToolNames(for:)` means "no restriction" — only
-        // non-nil sets participate in the intersection. Skills uses this
-        // to clamp the model's tool surface while a skill is active.
-        if let sessionRecord {
-            var allowList: Set<String>? = nil
-            for source in sessionToolSources {
-                if let allowed = await source.allowedToolNames(for: sessionRecord) {
-                    if let current = allowList {
-                        allowList = current.intersection(allowed)
-                    } else {
-                        allowList = allowed
-                    }
-                }
-            }
-            if let allowList {
-                unioned = unioned.filter { allowList.contains($0.name) }
-            }
-        }
-
-        // When the active backend declares a hard cap on how many tools it can
-        // handle per turn, truncate the list before building the GenerationConfig.
-        // The definitions are already sorted alphabetically by ToolRegistry, so
-        // `prefix` always picks the same lexicographically-first N tools —
-        // deterministic ordering keeps the "X of Y tools enabled" UI stable.
-        let cap = await MainActor.run { inferenceService.capabilities?.maxAdvertisedToolCount }
-        if let cap, unioned.count > cap {
-            return Array(unioned.prefix(cap))
-        }
-        return unioned
-    }
-
-    /// Registers a ``ToolExecutor`` adapter for each session-source-advertised
-    /// tool so a model tool call actually reaches the source's `resolve`
-    /// instead of coming back as ``ToolResult/ErrorKind/unknownTool`` (#1606).
-    ///
-    /// Returns the set of tool names this call registered so the caller can
-    /// unregister exactly those when the turn ends — leaving a session-scoped
-    /// executor in the shared registry would bind later turns to a stale
-    /// ``ChatSession``.
-    ///
-    /// A tool already present in the registry (a host-installed executor) is
-    /// left untouched: the registry executor wins, matching the
-    /// advertising-path de-dupe in ``readAdvertisedToolDefinitions``. Like the
-    /// per-turn `setHandoffDetector` / `setPreToolUseHook` wiring, this mutates
-    /// shared `InferenceService` state and assumes one turn per service at a
-    /// time; concurrent turns on the same service are last-writer-wins.
-    private func registerSessionToolExecutors(
-        sources: [any SessionToolSource],
-        sessionRecord: ChatSession?
-    ) async -> Set<String> {
-        guard let sessionRecord, !sources.isEmpty else { return [] }
-
-        // Gather (source, definition) pairs off the main actor so the
-        // per-source async `toolDefinitions(for:)` calls don't pin MainActor.
-        var pairs: [(source: any SessionToolSource, definition: ToolDefinition)] = []
-        for source in sources {
-            let defs = await source.toolDefinitions(for: sessionRecord)
-            for def in defs {
-                pairs.append((source, def))
-            }
-        }
-        guard !pairs.isEmpty else { return [] }
-
-        return await MainActor.run {
-            guard let registry = inferenceService.toolRegistry else { return Set<String>() }
-            var registered: Set<String> = []
-            for pair in pairs {
-                // Don't clobber a host-installed executor of the same name.
-                guard registry.contains(name: pair.definition.name) == false else { continue }
-                registry.register(
-                    SessionToolSourceExecutor(
-                        definition: pair.definition,
-                        source: pair.source,
-                        session: sessionRecord
-                    )
-                )
-                registered.insert(pair.definition.name)
-            }
-            return registered
-        }
-    }
-
-    private func unregisterSessionToolExecutors(_ names: Set<String>) async {
-        guard !names.isEmpty else { return }
-        await MainActor.run {
-            guard let registry = inferenceService.toolRegistry else { return }
-            for name in names {
-                registry.unregister(name: name)
-            }
-        }
     }
 
     @MainActor
